requirements.txt
snowflake-connector-python
pandas
psycopg2-binary
fastapi
uvicorn
pytest
python-dotenv
sqlalchemy
streamlit
plotly
kaggle